Moved teleop and added new auto
